Separating OfflinePlugin in many smaller dependencies - OpenClose principle
