+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/factory/OfflinePluginFactory.kt
@@ -1,0 +1,33 @@
+package io.getstream.chat.android.offline.experimental.plugin.factory
+
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
+import io.getstream.chat.android.offline.experimental.global.GlobalState
+import io.getstream.chat.android.offline.experimental.plugin.Config
+import io.getstream.chat.android.offline.experimental.plugin.OfflinePlugin
+import io.getstream.chat.android.offline.experimental.plugin.listener.ChannelMarkReadListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.EditMessageListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.GetMessageListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.HideChannelListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.MarkAllReadListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.QueryChannelListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.QueryChannelsListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.listener.ThreadQueryListenerImpl
+import io.getstream.chat.android.offline.experimental.plugin.logic.LogicRegistry
+
+@ExperimentalStreamChatApi
+internal class OfflinePluginFactory {
+
+    fun create(logic: LogicRegistry, globalState: GlobalState, config: Config): OfflinePlugin {
+        return OfflinePlugin(
+            queryChannelsListener = QueryChannelsListenerImpl(logic),
+            queryChannelListener = QueryChannelListenerImpl(logic),
+            threadQueryListener = ThreadQueryListenerImpl(logic),
+            channelMarkReadListener = ChannelMarkReadListenerImpl(logic),
+            editMessageListener = EditMessageListenerImpl(logic, globalState),
+            getMessageListener = GetMessageListenerImpl(logic),
+            hideChannelListener = HideChannelListenerImpl(logic),
+            markAllReadListener = MarkAllReadListenerImpl(logic),
+            config = config
+        )
+    }
+}
